patient-service/requirements.txt
Django
djangorestframework
psycopg2-binary
pika
